Extracting handleReset to ResetButton component
