+++ hang-man/src/App.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import './App.css';
-import generateRandomWord from './words'
+import generateRandomWord from './words';
 import gallow from './images/gallow.jpg';
 import KeyBoard from './components/KeyBoard';
 import Header from './components/Header';
@@ -9,9 +9,9 @@ import ResetButton from './components/ResetButton';
 function App (props) {
   
   //Hooks
-  let [count, updateCount] = useState(0);
-  let [randomWord, updateRandomWord] = useState(generateRandomWord());
-  let [guessedLetters, updateGuessedLetters] = useState([]);
+  let [count, setCount] = useState(0);
+  let [randomWord, setRandomWord] = useState(generateRandomWord());
+  let [guessedLetters, setGuessedLetters] = useState([]);
 
 
   const handleClick = e => {
@@ -20,18 +20,10 @@ function App (props) {
     console.log(element.getAttribute('key'));
     element.classList.add('letter-btn-inactive');
     element.disabled = true;
-    updateCount(++count);
+    setCount(++count);
   }
 
-  const handleReset = () => {
-    let letterButtons = Array.from(document.getElementsByClassName('letter-btn-inactive'));
-    letterButtons.map(letterButton => {
-      letterButton.classList.remove('letter-btn-inactive');
-      letterButton.disabled = false;
-    })
-    updateRandomWord(generateRandomWord());
-    updateCount(0);
-  }
+  
 
   const makeSecretWordRepresentation = () => {
     let secretWord = '';
@@ -47,10 +39,9 @@ function App (props) {
       <p id="secretWord" className="bigText">{makeSecretWordRepresentation()}</p>
       <p className="bigText">Antal gissningar: {count}</p>
       <div id="btn-container">
-      
         <KeyBoard className = 'letter-btn-default' letters={props.letters} handleClick = {handleClick}/>
       </div>
-      <ResetButton id= "reset-btn" handleReset={handleReset}/>
+      <ResetButton id= "reset-btn" setRandomWord = {setRandomWord} setCount={setCount}/>
     </div>
   );
 }
